backend/requirements.txt
absl-py==2.4.0
asgiref==3.11.1
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
Django==4.2.15
django-cors-headers==4.9.0
djangorestframework==3.16.1
filelock==3.24.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
jax==0.9.0.1
jaxlib==0.9.0.1
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mediapipe==0.10.9
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.2
opencv-contrib-python==4.13.0.92
opencv-python==4.8.1.78
opt_einsum==3.4.0
packaging==26.0
pillow==12.1.1
protobuf==3.20.3
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scipy==1.17.0
six==1.17.0
sounddevice==0.5.5
sqlparse==0.5.5
sympy==1.14.0
torch==2.10.0
typing_extensions==4.15.0
channels==4.1.0
daphne==4.1.2
channels-redis==4.2.0